+++ main_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/DATA-DOG/godog"
 	"github.com/DATA-DOG/godog/colors"
-	"github.com/optimizely/go-sdk/fscTest/support"
+	"github.com/optimizely/go-sdk/tests/integration/support"
 )
 
 var opt = godog.Options{Output: colors.Colored(os.Stdout)}
