+++ tests/integration/optimizely/listener/notification_listeners.go
@@ -1,7 +1,7 @@
 package listener
 
 import (
-	"github.com/optimizely/go-sdk/fscTest/optimizely/datamodels"
+	"github.com/optimizely/go-sdk/Tests/integration/optimizely/datamodels"
 	"github.com/optimizely/go-sdk/optimizely/decision"
 	"github.com/optimizely/go-sdk/optimizely/notification"
 )
